feat(docopt): actual docopt bindings
