+++ src/Cli.bs.js
@@ -1,0 +1,53 @@
+// Generated by BUCKLESCRIPT VERSION 4.0.5, PLEASE EDIT WITH CARE
+'use strict';
+
+var Block = require("bs-platform/lib/js/block.js");
+var Curry = require("bs-platform/lib/js/curry.js");
+var Docopt$ReasonRancherCli = require("./Docopt.bs.js");
+
+function readName(args) {
+  var match = args["<imageName>"];
+  if (match == null) {
+    return /* Service */Block.__(1, [args["<serviceName>"]]);
+  } else {
+    return /* Image */Block.__(0, [match]);
+  }
+}
+
+var D = Docopt$ReasonRancherCli.Parser(/* module */[]);
+
+function parse(str) {
+  var args = Curry._4(D[/* parse */0], undefined, undefined, undefined, str);
+  console.log(args);
+  if (args.upgrade) {
+    return /* Upgrade */Block.__(0, [
+              args["<stackName>"],
+              "pene"
+            ]);
+  } else if (args["upgrade-finish"]) {
+    return /* FinishUpgrade */Block.__(1, [
+              args["<stackName>"],
+              readName(args)
+            ]);
+  } else if (args.config) {
+    var match = args.saveEnv;
+    return /* Config */Block.__(3, [match ? /* SaveEnv */0 : /* Print */1]);
+  } else if (args["--version"]) {
+    return /* Version */0;
+  } else if (args.get) {
+    var match$1 = args.dockerCompose;
+    var match$2 = args["--output"];
+    return /* Get */Block.__(2, [
+              match$1 ? /* DockerCompose */0 : /* RancherCompose */1,
+              args["<stackName>"],
+              match$2 ? /* Name */[args["<fileName>"]] : /* NoFile */0
+            ]);
+  } else {
+    return /* Invalid */1;
+  }
+}
+
+exports.readName = readName;
+exports.D = D;
+exports.parse = parse;
+/* D Not a pure module */
