+++ src/Docopt.bs.js
@@ -1,49 +1,21 @@
 // Generated by BUCKLESCRIPT VERSION 4.0.5, PLEASE EDIT WITH CARE
 'use strict';
 
-var Block = require("bs-platform/lib/js/block.js");
 var Docopt = require("docopt");
 
-function readName(args) {
-  var match = args["<imageName>"];
-  if (match == null) {
-    return /* Service */Block.__(1, [args["<serviceName>"]]);
-  } else {
-    return /* Image */Block.__(0, [match]);
-  }
+function Parser() {
+  var parse = function ($staropt$star, $staropt$star$1, $staropt$star$2, doc) {
+    var help = $staropt$star !== undefined ? $staropt$star : true;
+    var options_first = $staropt$star$1 !== undefined ? $staropt$star$1 : false;
+    var exit = $staropt$star$2 !== undefined ? $staropt$star$2 : true;
+    return Docopt.docopt(doc, {
+                help: help,
+                options_first: options_first,
+                exit: exit
+              });
+  };
+  return /* module */[/* parse */parse];
 }
 
-function parse(str) {
-  var args = Docopt.docopt(str);
-  console.log(args);
-  if (args.upgrade) {
-    return /* Upgrade */Block.__(0, [
-              args["<stackName>"],
-              "pene"
-            ]);
-  } else if (args["upgrade-finish"]) {
-    return /* FinishUpgrade */Block.__(1, [
-              args["<stackName>"],
-              readName(args)
-            ]);
-  } else if (args.config) {
-    var match = args.saveEnv;
-    return /* Config */Block.__(3, [match ? /* SaveEnv */0 : /* Print */1]);
-  } else if (args["--version"]) {
-    return /* Version */0;
-  } else if (args.get) {
-    var match$1 = args.dockerCompose;
-    var match$2 = args["--output"];
-    return /* Get */Block.__(2, [
-              match$1 ? /* DockerCompose */0 : /* RancherCompose */1,
-              args["<stackName>"],
-              match$2 ? /* Name */[args["<fileName>"]] : /* NoFile */0
-            ]);
-  } else {
-    return /* Invalid */1;
-  }
-}
-
-exports.readName = readName;
-exports.parse = parse;
+exports.Parser = Parser;
 /* docopt Not a pure module */
